+++ tests/azure_mssql/test_azure_mssql.py
@@ -40,6 +40,6 @@ def test_gcmysql_get_df(mocker):
         user='ubuntu@localhost',
         database='mssql_db',
         password='ilovetoucan',
-        driver='{ODBC Driver 17 for SQL Server}',
+        driver='{ODBC Driver 18 for SQL Server}',
     )
     reasq.assert_called_once_with('my_query', con=snock(), params=[])