+++ toucan_connectors/install_scripts/mssql.sh
@@ -1,25 +1,25 @@
 #!/bin/bash
 set -e
 
+# Supports debian 12 and ubuntu 24.04
+
 if [[ -e ~/mssql-installed ]]; then
     echo "MSSQL connector dependencies are already installed."
     exit
 fi
 
 apt-get update
-apt-get install -fyq gnupg curl
-curl https://packages.microsoft.com/keys/microsoft.asc | apt-key add -
+apt-get install -fyq --no-install-recommends gnupg curl ca-certificates
+curl -fsSL https://packages.microsoft.com/keys/microsoft.asc -o /tmp/microsoft-key.asc
+# --batch to disable TTY and --yes to overwrite the file if it exists
+gpg --batch --yes --dearmor -o /usr/share/keyrings/microsoft-prod.gpg /tmp/microsoft-key.asc
+rm -f /tmp/microsoft-key.asc
 
 source /etc/os-release
-if [ "$ID" == "debian" ]; then
-    # debian/12 fails - fixing to debian/11 works:
-    curl "https://packages.microsoft.com/config/debian/11/prod.list" \
-        | tee /etc/apt/sources.list.d/mssql-release.list
-else
-    curl "https://packages.microsoft.com/config/${ID}/${VERSION_ID}/prod.list" \
-        | tee /etc/apt/sources.list.d/mssql-release.list
-fi
+curl "https://packages.microsoft.com/config/${ID}/${VERSION_ID}/prod.list" \
+    | tee /etc/apt/sources.list.d/mssql-release.list
+
 apt-get update
-ACCEPT_EULA=Y apt-get -y install msodbcsql17 unixodbc-dev
+ACCEPT_EULA=Y apt-get -y install msodbcsql18 unixodbc-dev
 
 touch ~/mssql-installed