+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Dependencies: loosened upper constraints
 
+### Changed
+
+- Loosened constraints on `elasticsearch` to allow using v9
+
 ## [8.0.2] 2025-04-14
 
 ### Fixed